created chared component Modal; used in delete modal and added modal to account page for creating new sdk key
